// src/api/types.gen.ts
export type Al = {
  readonly nadrukProperty?: Array<Nadruk> | null;
  readonly nadrukPropertySpecified?: boolean;
  readonly sup?: Array<string> | null;
  readonly supSpecified?: boolean;
  readonly inf?: Array<string> | null;
  readonly infSpecified?: boolean;
  readonly ovl?: Array<string> | null;
  readonly ovlSpecified?: boolean;
  readonly unl?: Array<string> | null;
  readonly unlSpecified?: boolean;
  readonly omissie?: Array<Omissie> | null;
  readonly omissieSpecified?: boolean;
  readonly nootref?: Array<Nootref> | null;
  readonly nootrefSpecified?: boolean;
  readonly noot?: Array<Noot> | null;
  readonly nootSpecified?: boolean;
  readonly aanhaling?: Array<Aanhaling> | null;
  readonly aanhalingSpecified?: boolean;
  readonly intref?: Array<Intref> | null;
  readonly intrefSpecified?: boolean;
  readonly extref?: Array<Extref> | null;
  readonly extrefSpecified?: boolean;
  readonly extrefGroep?: Array<ExtrefGroep> | null;
  readonly extrefGroepSpecified?: boolean;
  readonly intrefGroep?: Array<IntrefGroep> | null;
  readonly intrefGroepSpecified?: boolean;
  readonly dossierref?: Array<Dossierref> | null;
  readonly dossierrefSpecified?: boolean;
  readonly indexterm?: Array<Indexterm> | null;
  readonly indextermSpecified?: boolean;
  readonly afk?: Array<Afk> | null;
  readonly afkSpecified?: boolean;
  readonly cit?: Array<Cit> | null;
  readonly citSpecified?: boolean;
  readonly dfn?: Array<Dfn> | null;
  readonly dfnSpecified?: boolean;
  readonly redactie?: Array<Redactie> | null;
  readonly redactieSpecified?: boolean;
  readonly tekstcorrectie?: Array<Tekstcorrectie> | null;
  readonly tekstcorrectieSpecified?: boolean;
  text?: Array<string> | null;
  id?: string | null;
  lang?: Lang;
  langSpecified?: boolean;
};

export type AlGroep = {
  readonly al?: Array<Al> | null;
  readonly alSpecified?: boolean;
  readonly lijst?: Array<Lijst> | null;
  readonly lijstSpecified?: boolean;
};

export type AliasTitel = {
  status?: IAttlistAliasTitelStatus;
  text?: Array<string> | null;
};

export type AliasTitels = {
  readonly aliasTitel?: Array<AliasTitel> | null;
  readonly aliasTitelSpecified?: boolean;
  readonly afkorting?: Array<Afkorting> | null;
  readonly afkortingSpecified?: boolean;
};

export type Artikel = {
  kop: Kop;
  readonly lid?: Array<Lid> | null;
  readonly lidSpecified?: boolean;
  readonly structuurAlgemeen?: Array<unknown> | null;
  readonly structuurAlgemeenSpecified?: boolean;
  readonly table?: Array<Table> | null;
  readonly tableSpecified?: boolean;
  readonly wetcitaat?: Array<Wetcitaat> | null;
  readonly wetcitaatSpecified?: boolean;
  readonly binnenwet?: Array<Binnenwet> | null;
  readonly binnenwetSpecified?: boolean;
  artikelToelichting?: ArtikelToelichting;
  metaData?: MetaData;
  id?: string | null;
  lang?: Lang;
  langSpecified?: boolean;
  status?: ILokaalAttArtikelStatus;
  code?: string | null;
  type?: string | null;
  labelId?: string | null;
  stamId?: string | null;
  versieId?: string | null;
  publicatie?: IBwbcmsAttributenPublicatie;
  publicatieSpecified?: boolean;
  terugwerking?: string | null;
  variabelDeel?: string | null;
};
// ...
export type Boek = {
  kop: Kop;
  officieleInhoudsopgave?: OfficieleInhoudsopgave;
  structuurtekst?: Structuurtekst;
  readonly hoofdstuk?: Array<Hoofdstuk> | null;
  readonly hoofdstukSpecified?: boolean;
  readonly titeldeel?: Array<Titeldeel> | null;
  readonly titeldeelSpecified?: boolean;
  readonly afdeling?: Array<Afdeling> | null;
  readonly afdelingSpecified?: boolean;
  readonly paragraaf?: Array<Paragraaf> | null;
  readonly paragraafSpecified?: boolean;
  readonly artikel?: Array<Artikel> | null;
  readonly artikelSpecified?: boolean;
  readonly lokaalStructuurWetgeving?: Array<ClassWijzigArtikel> | null;
  readonly lokaalStructuurWetgevingSpecified?: boolean;
  metaData?: MetaData;
  id?: string | null;
  labelId?: string | null;
  stamId?: string | null;
  versieId?: string | null;
  publicatie?: IBwbcmsAttributenPublicatie;
  publicatieSpecified?: boolean;
  status?: IAttlistBoekStatus;
  statusSpecified?: boolean;
  terugwerking?: string | null;
};
// ...
export type Deel = {
  kop: Kop;
  officieleInhoudsopgave?: OfficieleInhoudsopgave;
  structuurtekst?: Structuurtekst;
  readonly hoofdstuk?: Array<Hoofdstuk> | null;
  readonly hoofdstukSpecified?: boolean;
  readonly titeldeel?: Array<Titeldeel> | null;
  readonly titeldeelSpecified?: boolean;
  readonly afdeling?: Array<Afdeling> | null;
  readonly afdelingSpecified?: boolean;
  readonly paragraaf?: Array<Paragraaf> | null;
  readonly paragraafSpecified?: boolean;
  readonly artikel?: Array<Artikel> | null;
  readonly artikelSpecified?: boolean;
  readonly lokaalStructuurWetgeving?: Array<ClassWijzigArtikel> | null;
  readonly lokaalStructuurWetgevingSpecified?: boolean;
  metaData?: MetaData;
  id?: string | null;
  labelId?: string | null;
  stamId?: string | null;
  versieId?: string | null;
  publicatie?: IBwbcmsAttributenPublicatie;
  publicatieSpecified?: boolean;
  status?: IAttlistDeelStatus;
  statusSpecified?: boolean;
  terugwerking?: string | null;
};
// ...
export type EnigArtikel = {
  readonly lid?: Array<Lid> | null;
  readonly lidSpecified?: boolean;
  readonly structuurAlgemeen?: Array<unknown> | null;
  readonly structuurAlgemeenSpecified?: boolean;
  readonly table?: Array<Table> | null;
  readonly tableSpecified?: boolean;
  readonly wetcitaat?: Array<Wetcitaat> | null;
  readonly wetcitaatSpecified?: boolean;
  readonly binnenwet?: Array<Binnenwet> | null;
  readonly binnenwetSpecified?: boolean;
  artikelToelichting?: ArtikelToelichting;
  metaData?: MetaData;
  id?: string | null;
  lang?: Lang;
  langSpecified?: boolean;
  status?: ILokaalAttArtikelStatus;
  code?: string | null;
  type?: string | null;
  labelId?: string | null;
  stamId?: string | null;
  versieId?: string | null;
  publicatie?: IBwbcmsAttributenPublicatie;
  publicatieSpecified?: boolean;
  terugwerking?: string | null;
};
// ...
export type Hoofdstuk = {
  kop: Kop;
  structuurtekst?: Structuurtekst;
  readonly titeldeel?: Array<Titeldeel> | null;
  readonly titeldeelSpecified?: boolean;
  readonly afdeling?: Array<Afdeling> | null;
  readonly afdelingSpecified?: boolean;
  readonly paragraaf?: Array<Paragraaf> | null;
  readonly paragraafSpecified?: boolean;
  readonly artikel?: Array<Artikel> | null;
  readonly artikelSpecified?: boolean;
  readonly lokaalStructuurWetgeving?: Array<ClassWijzigArtikel> | null;
  readonly lokaalStructuurWetgevingSpecified?: boolean;
  metaData?: MetaData;
  id?: string | null;
  status?: IAttlistHoofdstukStatus;
  statusSpecified?: boolean;
  terugwerking?: string | null;
  labelId?: string | null;
  stamId?: string | null;
  versieId?: string | null;
  publicatie?: IBwbcmsAttributenPublicatie;
  publicatieSpecified?: boolean;
};
// ...
export type IAttlistHoofdstukStatus = 0 | 1 | 2;

export type IAttlistIntituleStatus = 0 | 1 | 2;

export type IAttlistLidnrStatus = 0 | 1;

export type IAttlistLijstLevel = 0 | 1;

export type IAttlistNadrukType = 0 | 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8;

export type IAttlistNootLijstLevel = 0 | 1;

export type IAttlistOmissiePlaatsing = 0 | 1;

export type IAttlistParagraafStatus = 0 | 1 | 2;
// ...
export type IBwbcmsAttributenPublicatie = 0 | 1;

export type IContextAttType = 0 | 1 | 2 | 3 | 4 | 5 | 6 | 7;

export type ILokaalAttArtikelStatus = 0 | 1 | 2;
// ...
export type Kop = {
  readonly label?: Array<string> | null;
  readonly labelSpecified?: boolean;
  readonly nr?: Array<Nr> | null;
  readonly nrSpecified?: boolean;
  readonly titel?: Array<Titel> | null;
  readonly titelSpecified?: boolean;
  readonly subtitel?: Array<ClassTitel> | null;
  readonly subtitelSpecified?: boolean;
  kopopmaak?: KopKopopmaak;
  kopopmaakSpecified?: boolean;
};

export type KopKopopmaak = 0 | 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8;
// ...
export type Lijst = {
  readonly li: Array<Li>;
  type?: IAttLijstTypeType;
  id?: string | null;
  start?: string | null;
  level?: IAttlistLijstLevel;
  nrSluiting?: string | null;
  lang?: Lang;
  langSpecified?: boolean;
};
// ...
export type OfficieleInhoudsopgave = {
  kop?: Kop;
  readonly inhoud?: Array<Inhoud> | null;
  readonly inhoudSpecified?: boolean;
  inhoudGroep?: InhoudGroep;
  table?: Table;
  readonly al?: Array<Al> | null;
  readonly alSpecified?: boolean;
  metaData?: MetaData;
  id?: string | null;
  status?: ILokaalAttInhoudStatus;
  code?: string | null;
  type?: string | null;
  labelId?: string | null;
  stamId?: string | null;
  versieId?: string | null;
  publicatie?: IBwbcmsAttributenPublicatie;
  publicatieSpecified?: boolean;
  terugwerking?: string | null;
};
// ...
export type Paragraaf = {
  kop: Kop;
  structuurtekst?: Structuurtekst;
  readonly afdeling?: Array<Afdeling> | null;
  readonly afdelingSpecified?: boolean;
  readonly subParagraaf?: Array<SubParagraaf> | null;
  readonly subParagraafSpecified?: boolean;
  readonly artikel?: Array<Artikel> | null;
  readonly artikelSpecified?: boolean;
  readonly lokaalStructuurWetgeving?: Array<ClassWijzigArtikel> | null;
  readonly lokaalStructuurWetgevingSpecified?: boolean;
  metaData?: MetaData;
  id?: string | null;
  status?: IAttlistParagraafStatus;
  statusSpecified?: boolean;
  terugwerking?: string | null;
  labelId?: string | null;
  stamId?: string | null;
  versieId?: string | null;
  publicatie?: IBwbcmsAttributenPublicatie;
  publicatieSpecified?: boolean;
};
// ...
export type RegelingTekst = {
  enigArtikel?: EnigArtikel;
  readonly hoofdstuk?: Array<Hoofdstuk> | null;
  readonly hoofdstukSpecified?: boolean;
  readonly paragraaf?: Array<Paragraaf> | null;
  readonly paragraafSpecified?: boolean;
  readonly artikel?: Array<Artikel> | null;
  readonly artikelSpecified?: boolean;
  readonly wijzigArtikel?: Array<WijzigArtikel> | null;
  readonly wijzigArtikelSpecified?: boolean;
  readonly deel?: Array<Deel> | null;
  readonly deelSpecified?: boolean;
  tekst?: Tekst;
};
// ...
export type Tekst = {
  readonly wetcitaat?: Array<Wetcitaat> | null;
  readonly wetcitaatSpecified?: boolean;
  readonly alGroep?: Array<AlGroep> | null;
  readonly alGroepSpecified?: boolean;
  readonly structuurAlgemeen?: Array<unknown> | null;
  readonly structuurAlgemeenSpecified?: boolean;
  readonly table?: Array<Table> | null;
  readonly tableSpecified?: boolean;
  metaData?: MetaData;
  id?: string | null;
  lang?: Lang;
  langSpecified?: boolean;
  status?: ILokaalAttArtikelStatus;
  code?: string | null;
  type?: string | null;
  labelId?: string | null;
  stamId?: string | null;
  versieId?: string | null;
  publicatie?: IBwbcmsAttributenPublicatie;
  publicatieSpecified?: boolean;
  terugwerking?: string | null;
};
// ...
export type Wettekst = {
  officieleInhoudsopgave?: OfficieleInhoudsopgave;
  readonly al?: Array<Al> | null;
  readonly alSpecified?: boolean;
  readonly lijst?: Array<Lijst> | null;
  readonly lijstSpecified?: boolean;
  readonly alGroep?: Array<AlGroep> | null;
  readonly alGroepSpecified?: boolean;
  readonly boek?: Array<Boek> | null;
  readonly boekSpecified?: boolean;
  readonly deel?: Array<Deel> | null;
  readonly deelSpecified?: boolean;
  enigArtikel?: EnigArtikel;
  readonly hoofdstuk?: Array<Hoofdstuk> | null;
  readonly hoofdstukSpecified?: boolean;
  readonly titeldeel?: Array<Titeldeel> | null;
  readonly titeldeelSpecified?: boolean;
  readonly afdeling?: Array<Afdeling> | null;
  readonly afdelingSpecified?: boolean;
  readonly paragraaf?: Array<Paragraaf> | null;
  readonly paragraafSpecified?: boolean;
  readonly artikel?: Array<Artikel> | null;
  readonly artikelSpecified?: boolean;
  readonly lokaalStructuurWetgeving?: Array<ClassWijzigArtikel> | null;
  readonly lokaalStructuurWetgevingSpecified?: boolean;
};
// ...
export type WijzigArtikel = {
  kop: Kop;
  wat?: Wat;
  readonly al?: Array<Al> | null;
  readonly alSpecified?: boolean;
  readonly lijst?: Array<Lijst> | null;
  readonly lijstSpecified?: boolean;
  readonly wijzigLidGroep?: Array<WijzigLidGroep> | null;
  readonly wijzigLidGroepSpecified?: boolean;
  readonly wijzigLid?: Array<WijzigLid> | null;
  readonly wijzigLidSpecified?: boolean;
  readonly wijziging?: Array<Wijziging> | null;
  readonly wijzigingSpecified?: boolean;
  metaData?: MetaData;
  id?: string | null;
  status?: ILokaalAttArtikelStatus;
  code?: string | null;
  type?: string | null;
  labelId?: string | null;
  stamId?: string | null;
  versieId?: string | null;
  publicatie?: IBwbcmsAttributenPublicatie;
  publicatieSpecified?: boolean;
  terugwerking?: string | null;
};
// ...
export type WijzigLid = {
  lidnr: Lidnr;
  wat?: Wat;
  readonly wijziging?: Array<Wijziging> | null;
  readonly wijzigingSpecified?: boolean;
  id?: string | null;
};